requirements.txt
# FastAPI and dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.4.2
pydantic-settings>=2.0.3

# Database
motor>=3.3.1
pymongo>=4.6.0

# Environment variables
python-dotenv>=1.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
email-validator>=2.1.0
bcrypt>=4.0.1
